adding auth for dashboard and styles
