--- back/.env.example
@@ -1,8 +1,0 @@
-# Puerto del servidor
-PORT=3000
-# URL de MongoAtlas
-MONGODB_URI=mongodb+srv://<username>:<password>@<cluster-url
-# URL del frontend (para CORS)
-FRONTEND_URL=http://localhost:3000
-# Entorno de desarrollo
-NODE_ENV=development
