+++ back/src/controllers/authControllers.js
@@ -1,0 +1,72 @@
+import bcrypt from "bcrypt";
+import jwt from "jsonwebtoken";
+import User from "../models/User.js";
+
+
+export const registerUser = async (req, res) => {
+  const { email, password, nombre } = req.body;
+
+  if (!email || !password) {
+    return res.status(400).json({ error: "Email y contraseña son requeridos" });
+  }
+
+  try {
+    const existingUser = await User.findOne({ email });
+    if (existingUser) {
+      return res.status(400).json({ error: "El usuario ya existe" });
+    }
+
+    const hashedPassword = await bcrypt.hash(password, 10);
+
+    const user = new User({
+      email,
+      password: hashedPassword,
+      nombre
+    });
+
+    await user.save();
+
+    // Solo enviar mensaje sin token
+    res.status(201).json({ message: "Usuario creado correctamente" });
+
+  } catch (err) {
+    console.error("Error en registerUser:", err);
+    res.status(500).json({ error: "Error al registrar el usuario" });
+  }
+};
+
+
+export const loginUser = async (req, res) => {
+  const { email, password } = req.body;
+
+  if (!email || !password) {
+    return res.status(400).json({ error: "Email y contraseña son requeridos" });
+  }
+
+  try {
+    const user = await User.findOne({ email });
+    if (!user) {
+      return res.status(401).json({ error: "Credenciales inválidas" });
+    }
+
+    const passwordMatch = await bcrypt.compare(password, user.password);
+    if (!passwordMatch) {
+      return res.status(401).json({ error: "Credenciales inválidas" });
+    }
+
+    const token = jwt.sign(
+      {
+        id: user._id,
+        email: user.email,
+        rol: user.rol
+      },
+      process.env.JWT_SECRET,
+      { expiresIn: "2h" }
+    );
+
+    res.json({ token });
+  } catch (err) {
+    console.error("Error en loginUser:", err);
+    res.status(500).json({ error: "Error al iniciar sesión" });
+  }
+};
