+++ back/src/routes/authRoutes.js
@@ -1,0 +1,9 @@
+import express from "express";
+import { registerUser, loginUser } from "../controllers/authControllers.js";
+
+const router = express.Router();
+
+router.post("/register", registerUser); 
+router.post("/login", loginUser);
+
+export default router;
