+++ front/src/styles/Dashboard.css
@@ -1,0 +1,180 @@
+.dashboard-container {
+  display: flex;
+  min-height: 100vh;
+  font-family: "Poppins", "Segoe UI", sans-serif;
+  color: var(--text-color);
+
+  background: 
+    linear-gradient(135deg, #e0f0ff 0%, #f9fbff 100%),
+    url('https://www.transparenttextures.com/patterns/white-wall-3.png');
+  background-repeat: repeat;
+  background-size: auto;
+}
+
+
+/* Sidebar */
+.sidebar {
+  width: 280px;
+  background-color: var(--white);
+  box-shadow: 2px 0 8px rgba(0, 0, 0, 0.1);
+  padding: 2rem 1.5rem;
+  display: flex;
+  flex-direction: column;
+  position: sticky;
+  top: 0;
+  height: 100vh;
+}
+
+.sidebar-top {
+  text-align: center;
+  margin-bottom: 3rem;
+}
+
+.sidebar-logo {
+  width: 120px;
+  object-fit: contain;
+  margin-bottom: 0.5rem;
+  transition: var(--transition);
+  cursor: pointer;
+}
+
+.sidebar-logo:hover {
+  transform: scale(1.05);
+}
+
+.sidebar-appname {
+  font-size: 1.8rem;
+  font-weight: 700;
+  color: var(--primary-color);
+  margin-bottom: 0.3rem;
+}
+
+.sidebar-title {
+  font-size: 1.2rem;
+  color: var(--dark-color);
+  font-weight: 600;
+  border-bottom: 2px solid var(--primary-color);
+  padding-bottom: 0.3rem;
+  margin: 0 auto;
+  width: fit-content;
+  letter-spacing: 1px;
+}
+
+.sidebar-nav {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+.sidebar-link {
+  background: none;
+  border: none;
+  font-size: 1.1rem;
+  text-align: left;
+  padding: 0.8rem 1rem;
+  border-radius: 8px;
+  color: var(--text-color);
+  cursor: pointer;
+  font-weight: 500;
+  transition: var(--transition);
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.05);
+}
+
+.sidebar-link:hover,
+.sidebar-link.active {
+  background-color: var(--primary-color);
+  color: var(--white);
+  box-shadow: 0 4px 15px rgba(0, 123, 255, 0.3);
+  font-weight: 600;
+}
+
+/* Main Content */
+.main-content {
+  flex: 1;
+  padding: 3rem 2rem;
+}
+
+.dashboard-title {
+  font-size: 2rem;
+  color: var(--dark-color);
+  font-weight: 700;
+  margin-bottom: 1.5rem;
+  text-align: center;
+  position: relative;
+  padding-bottom: 1rem;
+  max-width: 600px;
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.table-wrapper {
+  max-width: 900px;
+  margin: 0 auto;
+  background: var(--white);
+  border-radius: 12px;
+  padding: 2rem;
+  box-shadow: 0 10px 30px rgba(0, 0, 0, 0.1);
+  overflow-x: auto;
+}
+
+.dashboard-table {
+  width: 100%;
+  border-collapse: collapse;
+}
+
+.dashboard-table thead tr {
+  background-color: var(--primary-color);
+  color: var(--white);
+}
+
+.dashboard-table th,
+.dashboard-table td {
+  padding: 1rem;
+  border: 1px solid #ddd;
+  text-align: left;
+}
+
+.dashboard-table tbody tr:hover {
+  background-color: #f0f8ff;
+  cursor: default;
+}
+
+.error {
+  color: #d32f2f;
+  font-weight: 600;
+  text-align: center;
+  margin-bottom: 1rem;
+}
+
+/* Responsive */
+@media (max-width: 768px) {
+  .dashboard-container {
+    flex-direction: column;
+  }
+  .sidebar {
+    width: 100%;
+    height: auto;
+    flex-direction: row;
+    padding: 1rem;
+    justify-content: space-between;
+  }
+  .sidebar-top {
+    margin-bottom: 0;
+    display: flex;
+    align-items: center;
+    gap: 1rem;
+  }
+  .sidebar-appname {
+    font-size: 1.4rem;
+  }
+  .sidebar-title {
+    display: none;
+  }
+  .sidebar-nav {
+    flex-direction: row;
+    gap: 1rem;
+  }
+  .main-content {
+    padding: 1.5rem 1rem;
+  }
+}
